requirements/dev.txt
-r base.txt

# To debug the app from the browser
# https://github.com/jazzband/django-debug-toolbar
django-debug-toolbar==3.2.1

# Debug and set breakpoints in VSCode
# https://github.com/microsoft/debugpy/
# TODO: get this to work
debugpy==1.5.1

# For creating mock data in test environments
# https://github.com/FactoryBoy/factory_boy
factory-boy==3.2.0

# For testing django with pytest
# https://pytest-django.readthedocs.io/en/latest/
pytest-django==4.4.0

# A better python shell
# https://ipython.org/
ipython==7.28.0

# Restart celery workers on file changes
# https://github.com/gorakhargosh/watchdog
watchdog[watchmedo]==2.1.6
